added carousel, hardcoded complications cards
