+++ client/package.json
@@ -24,6 +24,7 @@
     "passport-local": "^1.0.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-material-ui-carousel": "^1.8.3",
     "react-native": "^0.63.2",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
