+++ client/src/components/Complications/ComplicationsCarousel.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Carousel from 'react-material-ui-carousel'
+import { HeartCard, KidneyCard, EyeCard, NeuroCard, GumCard } from "./ComplicationsCards";
+import { Card } from '@material-ui/core';
+ 
+export default function ComplicationsCarousel() {
+    return (
+        <Carousel>
+            <HeartCard />
+            <KidneyCard />
+            <EyeCard />
+            <NeuroCard />
+            <GumCard />
+        </Carousel>
+    )
+}
