+++ client/src/pages/ResourcesPage.js
@@ -3,7 +3,7 @@ import CardGrid from "../components/CardGrid";
 import LowLevelCard from "../components/WarningsComponents/LowLevelCard";
 import HighLevelCard from "../components/WarningsComponents/HighLevelCard";
 import NormalLevelCard from "../components/WarningsComponents/NormalLevelCard";
-import ComplicationsCard from "../components/ComplicationsCard";
+import ComplicationsCarousel from "../components/Complications/ComplicationsCarousel";
 import CareScheduleAccordion from "../components/CareScheduleComponents/CareSchedule";
 import FoodTrackCard from "../components/FoodTrackCard";
 import Sources from "../components/Sources"
@@ -12,7 +12,7 @@ function ResourcesPage(props) {
   return (
     <CardGrid>
       <NormalLevelCard />
-      <ComplicationsCard />
+      <ComplicationsCarousel />
       <Sources />
     </CardGrid>
   );
